+++ examples/rgb-triangle.rs
@@ -3,7 +3,6 @@ extern crate quicksilver;
 
 use quicksilver::{
     run, Result, State,
-    geom::Vector,
     graphics::{Color, GpuTriangle, Vertex, Window, WindowBuilder}
 };
 
@@ -17,9 +16,9 @@ impl State for RgbTriangle {
     fn draw(&mut self, window: &mut Window) -> Result<()> {
         window.clear(Color::BLACK)?;
         let vertices = [
-            Vertex::new_untextured(Vector::new(400, 200), Color::RED),
-            Vertex::new_untextured(Vector::new(200, 400), Color::BLUE),
-            Vertex::new_untextured(Vector::new(600, 400), Color::GREEN),
+            Vertex::new_untextured((400, 200), Color::RED),
+            Vertex::new_untextured((200, 400), Color::BLUE),
+            Vertex::new_untextured((600, 400), Color::GREEN),
         ];
         let indices = [ GpuTriangle::new_untextured([0, 1, 2], 0.0) ];
         window.add_vertices(vertices.iter().cloned(), indices.iter().cloned());